requirement.txt
videosdk-agents>=0.0.64
videosdk-plugins-google>=0.0.64
videosdk-plugins-deepgram>=0.0.64
videosdk-plugins-cartesia>=0.0.64
videosdk-plugins-silero>=0.0.64
videosdk-plugins-turn-detector>=0.0.64
python-dotenv>=1.1.1
mcp>=1.0.0